- Implement a IKVStore Port and a KVStoreClient Redis adapter to be used
with the Mutex and AuthService (JWT implementation)

- Implement JWT tokens and Bearer authorization into existing
AuthService

+++ src/domains/Users/Service/UserService.ts
@@ -37,6 +37,8 @@ import { IPagingRequest } from '@src/domains/ports/persistence/IPagingRequest';
 import { IPagingResponse } from '@src/domains/ports/persistence/IPagingResponse';
 import { canNotBeEmpty, mustBePassword } from '@src/domains/validators';
 import { ServiceResponse } from '@src/infra/service/adapter/ServiceResponse';
+import { IMutexService } from '@src/infra/mutex/port/IMutexService';
+import { IPasswordCryptoService } from '@src/infra/security/IPasswordCryptoService';
 
 interface IUserServiceConfig extends IServiceConfig {
 
@@ -47,22 +49,31 @@ let userService: any;
 export class UserService extends BaseService<IUser, RequestCreateUser, RequestUpdateUser> {
   public dataRepository: UserDataRepository;
 
+  private entityName = 'User';
+
+  private mutexService: IMutexService;
+
+  private passwordCryptoService: IPasswordCryptoService;
+
   public constructor(
     config: IUserServiceConfig
   ) {
     super(config);
     const { dataRepository, services } = config;
     this.dataRepository = dataRepository as UserDataRepository;
-    this.services.passwordCryptoService = services!.passwordCryptoService;
+    this.passwordCryptoService = services!.passwordCryptoService;
+    // this.services.mutexService = services!.mutexService;
+    this.mutexService = services!.mutexService;
   }
 
   public async create(data: RequestCreateUser): Promise<IServiceResponse<IUser>> {
     const serviceResponse: IServiceResponse<IUser> = {};
     try {
-      mustBePassword('password', data.password);
+      const password = data.password || '';
+      mustBePassword('password', password);
 
       const newData = { ...data };
-      const { hash, salt } = await this.services.passwordCryptoService.hash(data.password);
+      const { hash, salt } = await this.passwordCryptoService.hash(password);
       newData.password = hash;
       newData.salt = salt;
 
@@ -78,10 +89,18 @@ export class UserService extends BaseService<IUser, RequestCreateUser, RequestUp
   public async update(id: string, data: RequestUpdateUser): Promise<IServiceResponse<IUser>> {
     const serviceResponse: IServiceResponse<IUser> = {};
     try {
+      const { result: { previouslyLocked } } = await this.mutexService.lock(this.entityName, id);
+      if (previouslyLocked) throw new Error(`${this.entityName} locked`);
+      // console.log('data', data);
       const user = await updateUser(id, data, this.dataRepository);
+      // console.log('user', user)
       serviceResponse.result = user;
+
+      await this.mutexService.unlock(this.entityName, id);
     } catch (error) {
       serviceResponse.error = error as Error;
+      // console.log(error);
+      await this.mutexService.unlock(this.entityName, id);
     }
     return serviceResponse;
   }
@@ -89,10 +108,17 @@ export class UserService extends BaseService<IUser, RequestCreateUser, RequestUp
   public async delete(id: string): Promise<IServiceResponse<boolean>> {
     const serviceResponse: IServiceResponse<boolean> = {};
     try {
+      const { result: { previouslyLocked } } = await this.mutexService.lock(this.entityName, id);
+      if (previouslyLocked) throw new Error(`${this.entityName} locked`);
+
       const deleted = await deleteUserById(id, this.dataRepository);
       serviceResponse.result = deleted;
+
+      await this.mutexService.unlock(this.entityName, id);
     } catch (error) {
       serviceResponse.error = error as Error;
+
+      await this.mutexService.unlock(this.entityName, id);
     }
     return serviceResponse;
   }
@@ -123,7 +149,6 @@ export class UserService extends BaseService<IUser, RequestCreateUser, RequestUp
     } catch (error) {
       serviceResponse.error = error as Error;
     }
-    // console.log(serviceResponse);
     return serviceResponse;
   }
 
@@ -136,14 +161,21 @@ export class UserService extends BaseService<IUser, RequestCreateUser, RequestUp
       canNotBeEmpty('password', data.password);
       mustBePassword('password', data.password);
 
+      const { result: { previouslyLocked } } = await this.mutexService.lock(this.entityName, id);
+      if (previouslyLocked) throw new Error(`${this.entityName} locked`);
+
       const newData = { ...data };
-      const { hash, salt } = await this.services.passwordCryptoService.hash(data.password);
+      const { hash, salt } = await this.passwordCryptoService.hash(data.password);
       newData.password = hash;
       newData.salt = salt;
-      const user = await updatePassword(id, data, this.dataRepository);
+      const user = await updatePassword(id, newData, this.dataRepository);
       serviceResponse.result = user;
+
+      await this.mutexService.unlock(this.entityName, id);
     } catch (error) {
       serviceResponse.error = error as Error;
+
+      await this.mutexService.unlock(this.entityName, id);
     }
     return new ServiceResponse(serviceResponse);
   }
@@ -154,39 +186,59 @@ export class UserService extends BaseService<IUser, RequestCreateUser, RequestUp
   ): Promise<IServiceResponse<IUser>> {
     const serviceResponse: IServiceResponse<IUser> = {};
     try {
+      const { result: { previouslyLocked } } = await this.mutexService.lock(this.entityName, id);
+      if (previouslyLocked) throw new Error(`${this.entityName} locked`);
+
       const user = await createDocument(id, data, this.dataRepository);
       serviceResponse.result = user;
+
+      await this.mutexService.unlock(this.entityName, id);
     } catch (error) {
       serviceResponse.error = error as Error;
+      await this.mutexService.unlock(this.entityName, id);
     }
     return serviceResponse;
   }
 
   public async updateDocument(
-    userId: string,
+    id: string,
     documentId: string,
     data: RequestUpdateDocument
   ): Promise<IServiceResponse<IUser>> {
     const serviceResponse: IServiceResponse<IUser> = {};
     try {
-      const user = await updateDocument(userId, documentId, data, this.dataRepository);
+      const { result: { previouslyLocked } } = await this.mutexService.lock(this.entityName, id);
+      if (previouslyLocked) throw new Error(`${this.entityName} locked`);
+
+      const user = await updateDocument(id, documentId, data, this.dataRepository);
       serviceResponse.result = user;
+
+      await this.mutexService.unlock(this.entityName, id);
     } catch (error) {
       serviceResponse.error = error as Error;
+
+      await this.mutexService.unlock(this.entityName, id);
     }
     return serviceResponse;
   }
 
   public async deleteDocument(
-    userId: string,
+    id: string,
     documentId: string
   ): Promise<IServiceResponse<IUser>> {
     const serviceResponse: IServiceResponse<IUser> = {};
     try {
-      const user = await deleteDocument(userId, documentId, this.dataRepository);
+      const { result: { previouslyLocked } } = await this.mutexService.lock(this.entityName, id);
+      if (previouslyLocked) throw new Error(`${this.entityName} locked`);
+
+      const user = await deleteDocument(id, documentId, this.dataRepository);
       serviceResponse.result = user;
+
+      await this.mutexService.unlock(this.entityName, id);
     } catch (error) {
       serviceResponse.error = error as Error;
+
+      await this.mutexService.unlock(this.entityName, id);
     }
     return serviceResponse;
   }
@@ -197,39 +249,60 @@ export class UserService extends BaseService<IUser, RequestCreateUser, RequestUp
   ): Promise<IServiceResponse<IUser>> {
     const serviceResponse: IServiceResponse<IUser> = {};
     try {
+      const { result: { previouslyLocked } } = await this.mutexService.lock(this.entityName, id);
+      if (previouslyLocked) throw new Error(`${this.entityName} locked`);
+
       const user = await createPhone(id, data, this.dataRepository);
       serviceResponse.result = user;
+
+      await this.mutexService.unlock(this.entityName, id);
     } catch (error) {
       serviceResponse.error = error as Error;
+
+      await this.mutexService.unlock(this.entityName, id);
     }
     return serviceResponse;
   }
 
   public async updatePhone(
-    userId: string,
+    id: string,
     phoneId: string,
     data: RequestUpdatePhone
   ): Promise<IServiceResponse<IUser>> {
     const serviceResponse: IServiceResponse<IUser> = {};
     try {
-      const user = await updatePhone(userId, phoneId, data, this.dataRepository);
+      const { result: { previouslyLocked } } = await this.mutexService.lock(this.entityName, id);
+      if (previouslyLocked) throw new Error(`${this.entityName} locked`);
+
+      const user = await updatePhone(id, phoneId, data, this.dataRepository);
       serviceResponse.result = user;
+
+      await this.mutexService.unlock(this.entityName, id);
     } catch (error) {
       serviceResponse.error = error as Error;
+
+      await this.mutexService.unlock(this.entityName, id);
     }
     return serviceResponse;
   }
 
   public async deletePhone(
-    userId: string,
+    id: string,
     phoneId: string
   ): Promise<IServiceResponse<IUser>> {
     const serviceResponse: IServiceResponse<IUser> = {};
     try {
-      const user = await deletePhone(userId, phoneId, this.dataRepository);
+      const { result: { previouslyLocked } } = await this.mutexService.lock(this.entityName, id);
+      if (previouslyLocked) throw new Error(`${this.entityName} locked`);
+
+      const user = await deletePhone(id, phoneId, this.dataRepository);
       serviceResponse.result = user;
+
+      await this.mutexService.unlock(this.entityName, id);
     } catch (error) {
       serviceResponse.error = error as Error;
+
+      await this.mutexService.unlock(this.entityName, id);
     }
     return serviceResponse;
   }
@@ -240,39 +313,60 @@ export class UserService extends BaseService<IUser, RequestCreateUser, RequestUp
   ): Promise<IServiceResponse<IUser>> {
     const serviceResponse: IServiceResponse<IUser> = {};
     try {
+      const { result: { previouslyLocked } } = await this.mutexService.lock(this.entityName, id);
+      if (previouslyLocked) throw new Error(`${this.entityName} locked`);
+
       const user = await createEmail(id, data, this.dataRepository);
       serviceResponse.result = user;
+
+      await this.mutexService.unlock(this.entityName, id);
     } catch (error) {
       serviceResponse.error = error as Error;
+
+      await this.mutexService.unlock(this.entityName, id);
     }
     return serviceResponse;
   }
 
   public async updateEmail(
-    userId: string,
+    id: string,
     emailId: string,
     data: RequestUpdateEmail
   ): Promise<IServiceResponse<IUser>> {
     const serviceResponse: IServiceResponse<IUser> = {};
     try {
-      const user = await updateEmail(userId, emailId, data, this.dataRepository);
+      const { result: { previouslyLocked } } = await this.mutexService.lock(this.entityName, id);
+      if (previouslyLocked) throw new Error(`${this.entityName} locked`);
+
+      const user = await updateEmail(id, emailId, data, this.dataRepository);
       serviceResponse.result = user;
+
+      await this.mutexService.unlock(this.entityName, id);
     } catch (error) {
       serviceResponse.error = error as Error;
+
+      await this.mutexService.unlock(this.entityName, id);
     }
     return serviceResponse;
   }
 
   public async deleteEmail(
-    userId: string,
+    id: string,
     emailId: string
   ): Promise<IServiceResponse<IUser>> {
     const serviceResponse: IServiceResponse<IUser> = {};
     try {
-      const user = await deleteEmail(userId, emailId, this.dataRepository);
+      const { result: { previouslyLocked } } = await this.mutexService.lock(this.entityName, id);
+      if (previouslyLocked) throw new Error(`${this.entityName} locked`);
+
+      const user = await deleteEmail(id, emailId, this.dataRepository);
       serviceResponse.result = user;
+
+      await this.mutexService.unlock(this.entityName, id);
     } catch (error) {
       serviceResponse.error = error as Error;
+
+      await this.mutexService.unlock(this.entityName, id);
     }
     return serviceResponse;
   }
